requirements.txt
flask>=3.0
numpy
rasterio
geopandas
shapely
scikit-image
pillow
pyogrio
matplotlib
huggingface_hub
gunicorn
